authentication, personal cabinet, orders
So, made it a lot of work. The main thing is auth, made it with redux and firebase, now possible to sign in and sign out. The second thing is a personal cabinet, which is a separate screen with tabs, where you can change info and check your orders(if you have it). Also changed the architecture of orders, now in order we have four separate nested objects(like delivery, credentials, etc.). Maybe later add some responsive design to the personal cabinet.